Refactor `generate_rnd_intervals` and add tests.
* Add `label` as an argument to `display_string`.
* Add tests for `display_string`.
* Refactor `generate_rnd_intervals` to more clearly handle different
  cases.
* Rename `generate_rnd_intervals` to `generate_intervals`.
* Add tests for `generate_intervals`.
* Add tests for `abs_diff_duration`.
* Have `generate_intervals` take an u8 instead of a usize for the
  number of intervals to generate, being in line with the maximum
  number of samples allowed by the program.

+++ src/main.rs
@@ -143,7 +143,7 @@ fn main() {
 
     match get_jitter(
         &args.host,
-        args.samples as usize,
+        args.samples,
         timeout,
         args.precision,
         args.min_interval,
